PolyUnion/compare: require the function to compare to be scalar
If we have overlaps, we need to do proper tiebreaking. For now we break
the ties based on the function to be compared (which is the right thing
when comparing cost functions anyhow).

We should think about a more general solution when comparing e.g.
feedback laws -- there we should use a different tiebreaking function.