Polyhedron/doesIntersect: fast detection of non-empty intersections
answer = P1.doesIntersect(P2) returns true if the intersection
of P1 and P2 is not an empty set.

answer = P1.doesIntersect(P2, 'fully') returns true if the
intersection is a fully dimensional set.

P1 and P2 must be single Polyhedron objects. The check
requires the H-representation of both polyhedrons.

This is a faster version of P1.intersect(P2).isEmptySet() and
P1.intersect(P2).isFulDim().

Note that this method is intended to be as fast as possible
for a good performance of PolyUnion/min()