even better options for gurobi
This is a follow-up to c129d984311a. Instead of turning the presolve
off globally in mpt_solvers_options(), we only prevent gurobi from
returning the fuzzy INF_OR_UNBD status locally in mpt_call_gurobi when
solving small-dimensional problems. Moreover, we achieve the goal by
setting InfUnbdInfo=1, which allows presolve to work for all problems.

As a consequence, we avoid having to resolve infeasible/unbounded problems
and still keep good performance of the solver itself.