Polyhedron/dual: computes the polar dual of a polytope
D = P.dual() computes the dual of a polyhedron P. The
polyhedron must be bounded (i.e., a polytope) and be fully
dimensional. The two exceptions are the dual of an empty set
(whose dual is R^n) and the dual of R^n (which is an empty
set). The input polytope is not required to contain the
origin in its interior.

The dual of P = { x | Ax \le b } (which contains the origin
in its interior) is D = { A'y | y \ge 0, \sum y=1}.

If the polytope is given in the vertex representation, i.e.,
P = { V'y | y \ge 0, \sum y = 1 }, then the polar dual is
D = { x | Vx \le 1 }.

fixes issue #97