New shortcut to create closed-loop systems
The overloaded '*' (mtimes) operator now creates an instance of the
ClosedLoop class, i.e., ctrl*sys is the same as ClosedLoop(ctrl, sys).

The mtimes operator is implemented both in AbstractSystem as well as in
AbstractController to allow for K*sys to work with K being the state
feedback gain matrix.