ULTISystem: class representing uncertain linear systems
To create an uncertain LTI system
  x^+ = A(L)x + B(L)u + E*d
    y = C(L)x + D(L)u
call:
  sys = ULTISystem('A', A, 'B', B, 'E', E, 'C', C, 'D', D)

where "A", "B", "C", "D" can be cell arrays of vertex
representations of the uncertain dynamics.

If "B", "C", "D" are not provided, they are
set to empty matrices. If "E" is not provided, E=eye(nx) is assumed.

Note that the additive disturbance is only considered if the
"E" matrix is set not a non-zero and non-empty matrix.

By default, the bounds on the disturbance are zero. These can be
changed via sys.d.min and sys.d.max.

More information can be found in Kvasnica, Takacs, Holaza, Ingole:
Reachability Analysis and Control Synthesis for Uncertain Linear Systems
in MPT; Submitted to ROCOND 2015.