Polyhedron: fullify the H- and V-representations
When A, b, Ae, be, H, He, V, and/or R are provided as sparse matrices,
they are now converted to full ones to prevent problems down the road.
There are simply too many places where we call rank(), svd(), and/or
various solvers which do not support sparse inputs. Catching this on
a per-case basis would be too costly.

Fixes issue #134, #135, #136
Might repoen issue #128