Opt/getRegion: compute the critical region and optimizers
pqp.getRegion(A) comptues the critical region, the parametric cost, and
the parametric optimizers (primal and dual) based on active inequalities
indexed in A.

Only supports pQPs without equalities for now.

In the long term, all parametric solvers should get the region just
via this function (requires more effort, especially on the side of
mpt_plcp).