Opt/getRegion: use the condensed primal optimizer in IPD descriptions
It turns out we do not need to store the full ([U; X]) primal optimizer
of the original problem with equalities. Instead, we store the optimizer
of the condensed QP. Note that this only influences region recovery and
set membership test. The "correct" primal optimizer is stored in
IPD.Functions('primal') anyhow.