IPDPolyhedron/isFullDim,isEmptySet - revert back to Polyhedron methods
It turns out our speed hack with fast_isEmptySet() and fast_isFullDim()
is not numerically robust and fails on a 10D case in test_enum_pqp_10.
Hence we are reverting back to using Polyhedron/isEmptySet() and
Polyhedron/isFullDim().

This change also requires an ugly hack to enable copying of trimmed
functions, as in test_enum_pqp_07 and _08.

Bottom line: premature optimization is the root of all evil!