IPDPolyUnion/toMatlab,toC: implement recovery of primals from duals
Primal variables can always be computed from the duals by
p=-inv(H)*(pF*x+f+G(as,:)'*d) once we know the optimal active set "as".

The behavior is controlled via the 'directPrimal' flag as in
	obj.toMatlab(fname, fun2export, 'directPrimal', true/false)
with "false" being the default=compute primals from duals.

Leads to smaller code as we do not have to export the parametric primal
optimizer for KKT checks.